refactor: 'HomeScreen' -> 'GroupSelectionScreen'
ce qui était désigné comme le homescreen est maintenant regroupé avec la partie planning étant donné
que celui-ci n'est plus le home. Cela permet une meilleure cohérence vis-à-vis de l'architecture de
la codebase.
